06_Dataengnner/sentEmailWithPython/pyproject.toml
[tool.poetry]
name = "sentemainwithpy"
version = "0.1.0"
description = ""
authors = ["Game <jeerasakananta@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
secure-smtplib = "0.1.1"
python-dotenv = "1.0.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
